lambda_function/requirements.txt
pylint
python-dateutil
boto3
morph
python-dateutil